requirements.txt
pandas==2.3.1
numpy==2.3.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
jupyterlab==4.4.4
matplotlib==3.10.3
seaborn